+++ tests/testthat/test-el_glm.R
@@ -18,13 +18,13 @@ test_that("Invalid `family`.", {
     family = poisson(link = make.link("inverse")),
     data = airquality
   ))
+  expect_error(el_glm(Temp ~ Wind,
+    family = quasipoisson(link = make.link("inverse")),
+    data = airquality
+  ))
   expect_error(el_glm(Wind ~ Temp,
     family = inverse.gaussian("identity"),
     data = airquality
-  ))
-  expect_error(el_glm(carb ~ .,
-    family = quasipoisson("sqrt"),
-    data = mtcars
   ))
 })